requirements.txt
agno>=0.1.0
turbopuffer>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.0.0